+++ NEWS.md
@@ -1,5 +1,7 @@
 # reuseme (development version)
 
+* `proj_outline()` and `dir_outline()` now excludes example files
+
 * `file_outline()` should recognize and transform markdown links automatically with new `markup_href()` It is no longer needed to use `{.href}` in your outline headings to show a link.
 
 * `link_issue()` has been renamed `link_gh_issue()` and now only takes care of changing gh issues in markdown links.